Separate single value tests to SingleValueTests.swift file
